+++ packages/sql-editor-config/src/SqlEditorSliceConfig.ts
@@ -13,12 +13,18 @@ export const SqlEditorSliceConfig = z.object({
     .string()
     .default('default')
     .describe('The id of the currently selected query.'),
+  lastExecutedQuery: z.string().optional().describe('Last executed query'),
+  closedTabIds: z
+    .array(z.string())
+    .default([])
+    .describe('List of closed tab ids'),
 });
 export type SqlEditorSliceConfig = z.infer<typeof SqlEditorSliceConfig>;
 
 export function createDefaultSqlEditorConfig(): SqlEditorSliceConfig {
   return {
-    queries: [{id: 'default', name: 'Untitled', query: ''}],
+    queries: [{id: 'default', name: 'SQL', query: ''}],
     selectedQueryId: 'default',
+    closedTabIds: [],
   };
 }